requirements.txt
python-dotenv
py-cord
pymongo